% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eclipse_env_data}
\alias{eclipse_env_data}
\title{Solar Eclipse Environment Sensor Data}
\format{
A data.table with 46,096 rows and 10 columns:
\describe{
\item{station_id}{Station ID}
\item{timestamp}{Reading timestamp with UTC offset (character)}
\item{temperature}{Temperature in Celsius}
\item{humidity}{Relative humidity (percent)}
\item{barometric_pressure}{Barometric pressure (hPa)}
\item{aqi}{Air quality index}
\item{eco2}{Equivalent CO2 (ppm)}
\item{voc}{Volatile organic compounds}
\item{sound_pressure_level}{Sound pressure level (dB)}
\item{datetime}{Parsed timestamp as POSIXct in local station time}
}
}
\source{
BirdWeather API via birdweatheR
}
\usage{
eclipse_env_data
}
\description{
BirdWeather PUC environmental sensor readings across stations in the path of
the April 8, 2024 total solar eclipse in North America.
}
\keyword{datasets}
