% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stl_env_May2025}
\alias{stl_env_May2025}
\title{St. Louis, Missouri USA Environment Sensor Data - May 2025}
\format{
A data.table with columns:
\describe{
\item{station_id}{Station ID}
\item{timestamp}{Reading timestamp}
\item{temperature}{Temperature in Celsius}
\item{humidity}{Relative humidity (percent)}
\item{barometric_pressure}{Barometric pressure (hPa)}
\item{aqi}{Air quality index}
\item{eco2}{Equivalent CO2 (ppm)}
\item{voc}{Volatile organic compounds}
\item{sound_pressure_level}{Sound pressure level (dB)}
}
}
\source{
BirdWeather API via birdweatheR
}
\usage{
stl_env_May2025
}
\description{
BirdWeather PUC environment sensor readings (temperature, humidity,
barometric pressure, AQI) in the St. Louis metro area surrounding the
May 15-16, 2025 tornado outbreak.
}
\keyword{datasets}
