% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stl_storm_May2025}
\alias{stl_storm_May2025}
\title{St. Louis, Missouri, USA Storm Detections - May 2025}
\format{
A data.table with detection and station columns as returned by
\code{get_detections()}
}
\source{
BirdWeather API via birdweatheR
}
\usage{
stl_storm_May2025
}
\description{
BirdWeather PUC detections in the St. Louis metro area surrounding the
May 15-16, 2025 tornado outbreak, including an EF3 tornado.
}
\keyword{datasets}
